First Commit
No art besides emotion icons, and even they are missing a few.

Characters can be selected, deselected and told where to go.  'Heroes' can target and begin combat with 'Monsters'.  Factions, base classes and stats are set up and work on a basic level.  Personalities and more are in but unused.  Some test sounds play, and base scene controller objects are present.

12:05am, 1/1/2018